Externalized scripts.js to multiple scripts
